When creating a song you can also add the artist automatically
+ Added CommandFactor which holds important MySqlCommands
* ExecuteScalar updated

+++ MySoundLib/Windows/MainWindow.xaml.cs
@@ -33,7 +33,7 @@ namespace MySoundLib.Windows
 
 			InitializeComponent();
 
-			var songExists = (long) _connectionManager.ExecuteScalar("select count(*) from songs");
+			var songExists = (int)_connectionManager.ExecuteScalar("select count(*) from songs");
 
 			if (songExists == 0)
 			{
